refactor: OmokDao를 제네릭 타입으로 수정
