+++ app/src/main/java/woowacourse/omok/data/OmokDataStoreImpl.kt
@@ -3,13 +3,14 @@ package woowacourse.omok.data
 import android.content.Context
 import woowacourse.omok.data.adapter.OmokEntityAdapter
 import woowacourse.omok.data.adapter.StonePosition
+import woowacourse.omok.data.`interface`.Dao
 import woowacourse.omok.data.`interface`.OmokDataStore
 import woowacourse.omok.model.board.Position
 import woowacourse.omok.model.board.Stone
 
 class OmokDataStoreImpl(
     context: Context,
-    private val omokDao: OmokDao = OmokDaoImpl(context)
+    private val omokDao: Dao<OmokEntity> = OmokDao(context)
 ): OmokDataStore {
     private val omoks = mutableListOf<OmokEntity>()
 
@@ -27,7 +28,7 @@ class OmokDataStoreImpl(
     }
 
     override fun lastStonePosition(): StonePosition? {
-        val omokEntity = omokDao.findLast()
+        val omokEntity = omokDao.findLastOrNull()
         return omokEntity?.run { OmokEntityAdapter.stonePosition(this) }
     }
