+++ app/src/main/java/woowacourse/omok/data/interface/Dao.kt
@@ -1,0 +1,15 @@
+package woowacourse.omok.data.`interface`
+
+interface Dao<T> {
+    fun save(entity: T): T
+
+    fun saveAll(entities: List<T>) {
+        entities.forEach { save(it) }
+    }
+
+    fun findAll(): List<T>
+
+    fun findLastOrNull(): T?
+
+    fun drop()
+}
